Fix python error

+++ core/amber/src/main/python/core/architecture/managers/statistics_manager.py
@@ -7,8 +7,8 @@ from proto.edu.uci.ics.amber.engine.architecture.worker import WorkerStatistics
 
 class StatisticsManager:
     def __init__(self):
-        self._input_tuple_count: Dict[int, long] = dict()
-        self._output_tuple_count: Dict[int, long] = dict()
+        self._input_tuple_count: Dict[int, int] = dict()
+        self._output_tuple_count: Dict[int, int] = dict()
         self._data_processing_time = 0
         self._control_processing_time = 0
         self._total_execution_time = 0
